Create Kurapika and Friends
Problem statement
Kurapika is long lost from his friends. He wants to meet his friends. He has total ‘n’ friends, and everyone is not available on the same day. But he can meet his friends only once. Given two arrays, ‘s’ and ‘e’, where s[i] denotes the starting date when an ith friend is available, and e[i] denotes when the ith friend will not be available after this date. Help Kurapika to know what is the maximum number of friends he can meet on one single day.

For Example :
s = {1, 4, 11} e = {9, 5, 13}

In the given example, the first friend is available between dates 1 to 9, the second is available on dates 4 and 5, and the third is available on dates 11 to 13. Hence he can meet a maximum of only two friends on dates 4 and 5.

Hence the answer is 2.
Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 <= T <= 100    
1 <= N <= 5000
1 <= s[i], e[i] <= 10^9

Time limit: 1 sec
Sample Input 1 :
2
3
1 5 7
2 6 8
7
1 2 3 4 5 6 7
8 9 10 11 12 13 14
Sample Output 1 :
1
7
Explanation For Sample Input 1 :
In the first test case, there was no single day when more than 1 friend was available.

Hence the answer is 1.

In the second test case, all of the friends were available on day 7 and day 8. 

Hence the final answer is 7.
Sample Input 2 :
2
5
1 2 6 5 3
5 5 7 6 8
4
1 2 3 4
10 3 3 4
Sample Output 2 :
4
3