Create Check If All Levels Of Two Trees Are Anagrams Of Each Other Not
Problem statement
You’re given two binary trees. Your task is to check if all the trees’ levels are anagrams of each other. If they’re anagram of each other, print ‘Yes’ else print ‘No.’

Note :

An Anagram is a word or phrase formed by rearranging the letters of a different word or phrase, typically using all the original letters exactly once.
Detailed explanation ( Input/output format, Notes, Images )
Sample Input 1:
1
3 3
1 2 3 -1 -1 -1 -1 
1 3 2 -1 -1 -1 -1
Sample Output 1:
Yes
Explanation of Sample Input 1:
The first level of both trees is an anagram of each other since it only has 1. The second level of the first tree contains elements 2, 3, which is also an anagram, to the second tree’s second level, including elements 3,2.

Sample Input 2 :
1
5 5
1 3 2 -1 -1 5 4 -1 -1 -1 -1 
1 2 3 4 5 -1 -1 -1 -1 -1 -1   
Sample Output 2 :
Yes
Explanation of Sample Input 2 :
The first level of both trees contains 1. Hence they are an anagram of each other. The second level of the first tree has 3,2, is an anagram to the second level of the second tree, which contains 2,3, and the third level of the first tree contains 5,4, which is also an anagram to the third level of the second tree, which includes 4,5.