Create Maximum Distance
Problem statement
You are given an array of binary strings ‘arr’ of size ‘N’. Your task is to find out the maximum distance between the pair of strings.

The distance between two binary strings is defined as the sum of their lengths after removing the common prefix.

For example:
You are given ‘arr’ = [ “01011”,  “010110”, “0111”], Here the strings with the maximum distance between them are “010110” and  “0111”, where the prefix is “01”,  hence the maximum distance is  4 + 2 = 6. Hence the answer is 6.
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 10
2 <= N <= 10^3
1 <= |arr[i]| <= 10^3

Time Limit: 1 sec
Note:
You do not need to print anything. It has already been taken care of. Just implement the function.
Sample Input 1:
2
1011100 1011011 1001111 0111111
01011 010110 0111 
Sample Output 1:
14
 6
Explanation:
For the first test case, ‘arr’ = ["1011100", "1011011", "1001111", "0111111"], Here the string with maximum distance between them are "1001111" and  "0111111”, where the prefix is “”,  hence the maximum distance is 7 + 7 = 14. Hence the answer is 14.

For the second test case, ‘arr’ = [ “01011”,  “010110”, “0111”], Here the strings with the maximum distance between them are “010110” and  “0111”, where the prefix is “01”. Hence the maximum distance is  4 + 2 = 6. Hence the answer is 6.
Sample Input 2:
2
1100 1110 0111 00
01000 01011 010 011
Sample Output 2:
8
4