Create Ninja And Sentences
Problem statement
Ninja is trying to create a robot, which would talk to humans. Ninja is stuck at a particular phase where he wants the robot to create all possible sentences from a 2D array of words. Help Ninja in solving this problem. For a given 2D array of words, your task is to create all possible sentences that can be made using them.

Note :
In a particular sentence, the robot can use a single word from the row that the word is part of.
Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 <= T <= 5
1 <= N <= 5
1 <= M <= 5
1 <= W <=20
where ‘T’ is the number of test cases,  ‘N’ and ‘M’ are the dimensions of the words array, ‘array[i][j]’ is the value of words of the array, and ‘W’ denotes the size of the word.
Sample Input 1:
1
2 2 
you we
sleep drink
Sample Output 1:
you sleep
you drink
we sleep
we drink
Explanation of Sample Input 1:
Taking a word from each row and adding it to each word of rest all the columns.
Sample Input 2:
2
3 1
i
love
coding
3 2
you we
have are
eat read
Sample Output 2:
i love coding
you have eat
you have read
you are eat
you are read
we have eat
we have read
we are eat
we are read