Create Inform Employees
Problem statement
You are in a company with ‘N’ employees each with a unique ID from ‘0 to N - 1’. You are given a list of managers, where ‘manager[i]’ is the direct manager of ‘ith’ employee. You are also given an array ‘timeToInform’ where ‘timeToInform[i]’ is the time it takes for ‘ith’ employee to send information to all his subordinates. Your task is to find out how much time it will take for any piece of information to reach all employees. You are also given ‘headId’ that is the ID of the head of the company.

For example:
You are given ‘manager’ = [-1, 0, 0, 1, 1], ‘timeToInform’ = [1, 1, 0, 0, 0], ‘headId’ = [0], 

We can see employee ‘0’ will take 1 time to inform their subordinates, that are employee 1, and employee 2, and employee 1 will take 1 time to inform their subordinates that are employee 3 and employee 4. The total time taken is 2. Hence the answer is 2.
Detailed explanation ( Input/output format, Notes, Images )
Constraints
1 <= T <= 10
1 <= N <= 10^5
0 <= headId, manager[i] < N
0 <= timeToInform[i] <= 10^4
manager[headId] = -1

Time Limit = 1 sec
Sample Input 1
2
5 0
-1 0 0 1 1
1 1 0 0 0
3 1
1 -1 0
1 2 0
Sample Output 1:
2
3
Explanation:
For the first test case, ‘manager’ = [-1, 0, 0, 1, 1], ‘timeToInform’ = [1, 1, 0, 0, 0], ‘headId’ = [0], 

We can see employee ‘0’ will take 1 time to inform their subordinates, that are employee 1, and employee 2, and employee 1 will take 1 time to inform their subordinates that are employee 3 and employee 4. The total time taken is 2. Hence the answer is 2.


For the second test case, ‘manager’ = [1, -1, 0], ‘timeToInform’ = [1, 2, 0], ‘headId’ = 1, 

We can see employee ‘1’ will take 2 time to inform their subordinates, that is employee ‘0’ and employee ‘0’ will take 1 time to inform their subordinate, i.e, employee 2. Hence the total time taken is 3
Sample Input 2:
2
1 0
-1
0
3 0
-1 0 0
2 0 0
Sample Ouput 2:
0
2