Create NINJA AND HIS GIRLFRIEND
Problem statement
Ninja and his girlfriend want to shift to a new locality and want their houses distant. The homes in that locality are like nodes of a “Binary Search Tree.”

The distance between the houses is the difference between the house numbers.

Help Ninja and his girlfriend find the required location.

So, your task is to find the minimum absolute difference between any two nodes.

Note:

A binary search tree is a binary tree data structure, with the following properties
The left subtree of any node contains nodes with a value less than the node’s value. The right subtree of any node contains nodes with a value equal to or greater than the node’s value. Right, and left subtrees are also binary search trees.

Example:

consider the following binary search tree, so the minimum absolute difference is between the node value, i.e., | 15-13 |= ‘2’so ‘2’is our answer.

Detailed explanation ( Input/output format, Notes, Images )
Sample Input 1:
2
5 3 7 2 4 6 8 -1 -1 -1 -1 -1 -1 -1 -1
1 -1 6 -1 -1
Sample Output 1:
1
5
Explanation for Sample 1:
Test Case 1:
The level order traversal of a given tree is 5 3 7 2 4 6 8 -1 -1 -1 -1 -1 -1 -1 -1; thus, its first root is root number with value 5 and left, and right nodes of the 5 are nodes with 3 and 7 respectively.
The binary tree according to the test case is:


So we choose the node which has a minimum absolute difference, i.e., ‘1’, so our answer is ‘1’.

Test Case 2:
The binary tree according to the test case is:

So as there are only two nodes in this tree, so on taking their absolute minimum difference, i.e ‘6 - 1’  we get ‘5’, so our answer is ‘5’.
Sample Input2:
 2
2 1 -1 -1 -1
10 -1 20 -1 30 -1 40 -1 -1
Sample Output2:
1
10