Create Moving Average
Problem statement
A stock price is dynamic. Its value can change multiple times in a fraction of a second or remain unchanged for several minutes. Analysing the dynamics of stock price change can provide an indication of a forthcoming uptrend or downtrend in that stock. One such indicator is simple moving averages. Now, Harry wants to analyse the price trend of the stock on the basis of moving averages (MA).

Let’s consider a moving average of 2-day and 4-day, respectively. A 2-day moving average is calculated by taking an average of the closing price of 2 consecutive days. A 4-day moving average is calculated by taking an average closing price of 4 consecutive days. Now, according to experts whenever a faster moving average curve (2-day MA) cuts the slower moving average (4-day MA) from below, then it is an indication of an uptrend in the stock. Similarly, whenever a faster-moving averages curve (2-day MA) cuts the slower moving average curve (4-day MA) from above, then it is an indication of the downtrend in the stock.

Help Harry in computing the number of uptrends and downtrends in the given time for which the data is provided.


In this graph, there are three lines indicating stock closing price, moving average of two days and four days. Now we can see that between 13th and 15th there is an intersection. It is known as a downtrend when the moving average of fewer days is cut downwards the more days moving average and vice versa.

Note1 – There will be no day1 moving average for 2-day MA. Similarly, there will be no day1, day2, and day3 moving average for 4-day MA. In general, there will be no X-1, X-2, Y-1, Y-2, etc. day point for the X-day and Y-day moving average curve.

Note2 – All the computation has to be accurate up to 6 digits after the decimal point.

EXAMPLE:
Input: 'X' = 2, 'Y' = 3, 'N' = 4, A = [2.0, 4.0, 3.0, 2.0]

Output: "1"

As moving average of length ‘2’ till index ‘2’ is ‘((4.0+3.5)/2)’=’3.5’, and the moving average of length ‘3’ till index ‘2’ is  ‘((2.0+4.0+3.0)/3)’=’3.0’. As we move to the next index ‘3’, the moving average of length ‘2’ becomes ‘2.5’ and of length ‘3’ becomes ‘3.0’. So moving average to length ‘2’ changed smaller at current index from bigger at last index from moving average of length ‘3’, so they are cutting each other between this period.
Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 <= 'T' <= 10
1 <= ‘N’ <= 10^5
1 <= 'X', ‘Y’ <= 10^5
1.0 <= ‘A[i]’ <= 10^4
Time Limit: 1 sec
Sample Input 1 :
2
2 3
4
5.0 1.0 3.0 2.0
3 5
11
4.55 5.4 5.65 5.4 5.2 4.85 4.95 5.05 4.9 4.9 4.95
Sample Output 1 :
1
3
Explanation Of Sample Input 1 :
In test case ‘1’. As moving average of length ‘2’ till index ‘2’ is ‘((1.0+3.0)/2)’=’2.0’, and an average of length ‘3’ till index ‘2’ is  ‘((5.0+1.0+3.0)/3)’=’3.0’. 

As we move to the next index ‘3’, the moving average of length ‘2’ becomes ‘2.5’, and of length, ‘3’ becomes ‘2.0’. 

So moving average to length ‘2’ has changed to bigger at current index from smaller at last index so they are cutting each other between this period. So the answer is incremented by ‘1’. 

In test case ‘2’. As moving average of length ‘3’ at index ‘4’ is ‘5.416666’ and the average of length ‘5’ is ‘5.24’.

As we move to the next index ‘5’ average of length ‘3’ changes to ‘5.15’ and of length ‘5’ changes to  ‘5.3’. So moving average of length ‘3’ changes from bigger to smaller at current index so answer increment by ‘1’.and as we will move ahead it will cut the slow-moving average ‘2’ time more so the final answer will be ‘3’.
Sample Input 2 :
2
2 4
14
69.849998 72.900002 74.449997 77.300003 75.050003 74.349998 75.449997 76.300003 74 69.349998 65.349998 67.349998 67.599998 68.449997
3 3
4
1.2 2.4 3.6 4.8
Sample Output 2 :
4
0